requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow==9.4.0
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.8.0
torchvision>=0.9.0
tqdm>=4.41.0

# logging -------------------------------------
tensorboard>=2.4.1
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.9.1
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop  # FLOPS computation
pycocotools>=2.0  # COCO mAP

# Flask
Flask==2.2.3
Flask-Login==0.6.2
python-dotenv==1.0.0 #抓.flaskenv環境變數用的 
flask_mysqldb==1.0.1 
mysqlclient==2.1.1



## pip freeze > requirements.txt 出來的結果

# certifi==2022.12.7
# charset-normalizer==3.1.0
# click==8.1.3
# colorama==0.4.6
# contourpy==1.0.7
# cycler==0.11.0
# Flask==2.2.3
# fonttools==4.39.0
# idna==3.4
# importlib-metadata==6.0.0
# importlib-resources==5.12.0
# itsdangerous==2.1.2
# Jinja2==3.1.2
# kiwisolver==1.4.4
# MarkupSafe==2.1.2
# matplotlib==3.7.1
# numpy==1.24.2
# opencv-python==4.7.0.72
# packaging==23.0
# pandas==1.5.3
# Pillow==9.4.0
# pyparsing==3.0.9
# python-dateutil==2.8.2
# python-dotenv==1.0.0
# pytz==2022.7.1
# PyYAML==6.0
# requests==2.28.2
# scipy==1.10.1
# seaborn==0.12.2
# six==1.16.0
# torch==1.8.0
# torchaudio==0.13.1+cu116
# torchvision==0.9.0
# tqdm==4.65.0
# typing_extensions==4.5.0
# urllib3==1.26.14
# Werkzeug==2.2.3
# zipp==3.15.0
